+++ build.gradle
@@ -17,8 +17,11 @@ dependencies {
     implementation group: 'org.apache.commons', name: 'commons-collections4', version: '4.4'
     implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.13.0'
 
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
+    implementation group: 'org.mockito', name: 'mockito-junit-jupiter', version: '5.11.0'
+    testCompileOnly 'junit:junit:4.12'
+    implementation("org.assertj:assertj-core:3.24.2")
+    implementation 'org.junit.jupiter:junit-jupiter-api:5.10.0'
+    implementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
 }
 
 test {
