+++ src/main/java/firstTask/com/exceptions/NotUniqueUserNameException.java
@@ -1,0 +1,16 @@
+package firstTask.com.exceptions;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+
+/**
+ * Класс ответственен за обязательную обработку события неуникальности пользовательского имени
+ * @author  Zherko Vladislav
+ * @since   1.0
+ */
+@AllArgsConstructor
+@Getter
+public class NotUniqueUserNameException extends Exception {
+    private String message;
+}
