+++ src/main/java/firstTask/com/in/TrainingDiaryApplication.java
@@ -1,19 +1,32 @@
-package firstTast.com.in;
+package firstTask.com.in;
 
-import firstTast.com.exceptions.NotUniqueUserNameException;
-import firstTast.com.exceptions.NotUniqueWorkoutTypeException;
-import firstTast.com.model.ConsoleUser;
-import firstTast.com.service.AuthenticationService;
-import firstTast.com.service.UserActionService;
-import firstTast.com.service.impl.WorkoutUpdateServiceImpl;
-import firstTast.com.util.MenuOptions;
-import firstTast.com.util.UtilScanner;
+import firstTask.com.model.ConsoleUser;
+import firstTask.com.util.MenuOptions;
+import firstTask.com.util.UtilScanner;
+import firstTask.com.exceptions.NotUniqueUserNameException;
+import firstTask.com.exceptions.NotUniqueWorkoutTypeException;
+import firstTask.com.model.Workout;
+import firstTask.com.service.AuthenticationService;
+import firstTask.com.service.UserActionService;
+import firstTask.com.service.impl.WorkoutUpdateServiceImpl;
 import lombok.*;
 
+import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.Optional;
 import java.util.Scanner;
 import java.util.logging.Logger;
+
+
+/**
+ *
+ * Класс, ответственный за контакт и обратную связь с пользователем (консолью):
+ *
+ * @see AuthenticationService - работа с аунтификацией
+ * @see UserActionService - обработка действий пользователя
+ * @see WorkoutUpdateServiceImpl - обработка действий, связанных непосредственно с задачами
+ * @see ConsoleUser - пользователь приложения
+ *  **/
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -25,6 +38,9 @@ public class TrainingDiaryApplication {
     private WorkoutUpdateServiceImpl workoutService;
     private ConsoleUser consoleUser;
 
+    /**
+     * Метод, запускающий приложение
+     *  **/
     public void startApplication() {
         while (true) {
             if(!enterApplication()){
@@ -38,7 +54,7 @@ public class TrainingDiaryApplication {
 
     private boolean enterApplication() {
         do {
-            Scanner scanner = UtilScanner.getScanner();
+            Scanner scanner = UtilScanner.getInstance();
             System.out.println(MenuOptions.enterMenu);
             String input = scanner.nextLine();
 
@@ -82,7 +98,8 @@ public class TrainingDiaryApplication {
         switch (menuChoice) {
             case "1", "Добавить тренировку" -> {
                 try {
-                    System.out.println(userService.addNewWorkout(consoleUser).getType() + " - тренировка добавлена");
+                    Workout workout = createWorkout();
+                    System.out.println(userService.addNewWorkout(consoleUser, workout).getType() + " - тренировка добавлена");
                     logger.info(consoleUser.getUsername() + " добавление тренировки: SUCCESS");
                 } catch (DateTimeParseException e) {
                     System.out.println("Неверна введена дата");
@@ -96,11 +113,10 @@ public class TrainingDiaryApplication {
                 changeWorkout();
             }
             case "3", "Просмотреть тренировки" -> {
-                System.out.println(userService.showAllWorkoutsDateSorted(consoleUser));
+                System.out.println(userService.showAllWorkoutsDateSorted(consoleUser).toString());
             }
             case "4", "Статистика тренировок" -> {
                 System.out.println(userService.getWorkoutStatistics(consoleUser));
-                logger.info(consoleUser.getUsername() + " получение статистики: SUCCESS");
             }
             case "5" -> {
                 if (consoleUser.getRole().equals("ADMIN")) {
@@ -116,17 +132,43 @@ public class TrainingDiaryApplication {
             }
         }
     }
+    private Workout createWorkout(){
+        System.out.println("Введите дату вашей тренировки в виде YYYY-MM-DD : ");
+        Scanner scanner = UtilScanner.getInstance();
+        String date = scanner.nextLine();
+        LocalDate localDate = LocalDate.parse(date);
+
+        System.out.println("Введите тип тренировки : ");
+        String type = scanner.nextLine();
+
+        System.out.println("Введите продолжительность тренировки : ");
+        double time = scanner.nextDouble();
+
+        System.out.println("Введите кол-во сожженных калорий : ");
+        double calories = scanner.nextDouble();
+
+        System.out.println("Опишите свою тренировку : ");
+        scanner.nextLine();
+        String additional = scanner.nextLine();
+
+        return Workout.builder()
+                .timeOfWorkout(localDate)
+                .type(type)
+                .minuteDuration(time)
+                .caloriesBurned(calories)
+                .additionalInfo(additional)
+                .build();
+    }
     private void userAction() {
         String menuChoice = "";
         do {
             System.out.println(MenuOptions.userMenu);
-            menuChoice = UtilScanner.getScanner().nextLine();
+            menuChoice = UtilScanner.getInstance().nextLine();
             handleMenuChoice(menuChoice, consoleUser);
-        } while (!menuChoice.equals("7") && !menuChoice.equals("Назад"));
+        } while (!menuChoice.equals("6") && !menuChoice.equals("Назад"));
     }
-
     private void changeWorkout() {
-        Scanner scanner = UtilScanner.getScanner();
+        Scanner scanner = UtilScanner.getInstance();
         Integer workoutByIndex = workoutService.getWorkoutByIndex(consoleUser);
 
         if (workoutByIndex < 0) {
@@ -147,13 +189,17 @@ public class TrainingDiaryApplication {
                 System.out.print("Введите новую дату тренировки : ");
                 String newDate = scanner.nextLine();
                 workoutService.changeDate(consoleUser, newDate);
-                logger.info(consoleUser.getUsername() + " новая дату тренировки: SUCCESS");
             }
             case "3" -> {
                 System.out.print("Введите новую продолжительность тренировки : ");
                 Double newTime = scanner.nextDouble();
-                workoutService.changeMinuteDuration(consoleUser, newTime);
-                logger.info(consoleUser.getUsername() + " новая продолжительность тренировки: SUCCESS");
+                try {
+                    workoutService.changeMinuteDuration(consoleUser, newTime);
+                    logger.info(consoleUser.getUsername() + " новая продолжительность тренировки: SUCCESS");
+                } catch (DateTimeParseException e) {
+                    System.out.println("Некорректный ввод даты");
+                    logger.info(consoleUser.getUsername() + " новая продолжительность тренировки: FAIL");
+                }
             }
             case "4" -> {
                 System.out.print("Введите новые калории : ");
