+++ src/main/java/firstTask/com/util/MenuOptions.java
@@ -1,12 +1,20 @@
-package firstTast.com.util;
+package firstTask.com.util;
 
+/**
+ * Класс, содержащий все меню-опции приложения
+ */
 public class MenuOptions {
+    /**
+     * Начальная меню-опция
+     */
     public static String enterMenu = """
             1. Вход
             2. Регистрация
             3. Завершить программу
             """;
-
+    /**
+     * Меню-опция для управления тренировками
+     */
     public static String userMenu = """
             1. Добавить тренировку
             2. Редактировать тренировку
@@ -16,6 +24,9 @@ public class MenuOptions {
             6. Назад
             """;
 
+    /**
+     * Меню-опция для изменения выбранной тренировки
+     */
     public static String workoutEditorMenu = """
             1. Новое название
             2. Новая дата
