--- src/main/java/firstTast/com/exceptions/NotUniqueUserNameException.java
@@ -1,10 +1,0 @@
-package firstTast.com.exceptions;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class NotUniqueUserNameException extends Exception {
-    private String message;
-}
