--- src/main/java/firstTast/com/exceptions/NotUniqueWorkoutTypeException.java
@@ -1,10 +1,0 @@
-package firstTast.com.exceptions;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class NotUniqueWorkoutTypeException extends Throwable {
-    private String message;
-}
