--- src/main/java/firstTast/com/util/UtilScanner.java
@@ -1,9 +1,0 @@
-package firstTast.com.util;
-
-import java.util.Scanner;
-
-public class UtilScanner {
-    public static Scanner getScanner(){
-        return new Scanner(System.in);
-    }
-}
